# apps/api/src/workspaces/webhooks.controller.ts
import {
  Controller,
  Get,
  Post,
  Patch,
  Delete,
  Body,
  Param,
  UseGuards,
  BadRequestException,
  NotFoundException,
  ForbiddenException,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse, ApiBearerAuth, ApiParam, ApiBody, ApiProperty } from '@nestjs/swagger';
import { AuthGuard } from '../auth/auth.guard';
import { CurrentUser } from '../auth/current-user.decorator';
import { prisma } from '@repo/database';
import type { User } from '@repo/database';
import { z } from 'zod';
import * as crypto from 'crypto';
import { IsString, IsArray, IsUrl, IsOptional, IsBoolean } from 'class-validator';

class CreateWorkspaceWebhookDto {
  @IsString()
  @ApiProperty({ example: 'My Webhook' })
  name: string;

  @IsUrl()
  @ApiProperty({ example: 'https://example.com/webhook' })
  url: string;

  @IsArray()
  @IsString({ each: true })
  @ApiProperty({ type: [String], example: ['message.sent', 'channel.created'] })
  events: string[];
}

class UpdateWorkspaceWebhookDto {
  @IsBoolean()
  @IsOptional()
  @ApiProperty({ required: false, example: true })
  active?: boolean;
}

const createWebhookSchema = z.object({
  name: z.string().min(1),
  url: z.string().url(),
  events: z.array(z.string()),
});

const updateWebhookSchema = z.object({
  active: z.boolean().optional(),
});

@ApiTags('Webhooks')
@ApiBearerAuth()
@Controller('workspaces/:slug/webhooks')
@UseGuards(AuthGuard)
export class WebhooksController {
  @Get()
  @ApiOperation({ summary: 'Get all webhooks for a workspace' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiResponse({ status: 200, description: 'List of webhooks' })
  async getWebhooks(@CurrentUser() user: User, @Param('slug') slug: string) {
    /**
     * ⚡ Performance Optimization:
     * 1. Consolidates workspace lookup, membership verification, and webhook retrieval into a single query.
     * 2. Uses nested 'select' to fetch only required fields and relations (like log counts).
     * 3. Reduces database round-trips from 2 down to 1.
     * Expected impact: Faster response times and improved security with explicit membership check.
     */
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
        webhooks: {
          include: {
            _count: {
              select: {
                logs: true,
              },
            },
          },
          orderBy: { createdAt: 'desc' },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    if (workspace.members.length === 0) {
      throw new ForbiddenException('You do not have access to this workspace');
    }

    return workspace.webhooks;
  }

  @Post()
  @ApiOperation({ summary: 'Create a new webhook for a workspace' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiBody({ type: CreateWorkspaceWebhookDto })
  @ApiResponse({ status: 201, description: 'Webhook created successfully' })
  async createWebhook(@CurrentUser() user: User, @Param('slug') slug: string, @Body() body: CreateWorkspaceWebhookDto) {
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    const member = workspace.members[0];

    if (!member || !['owner', 'admin'].includes(member.role)) {
      throw new ForbiddenException('Forbidden: Only owners and admins can create webhooks');
    }

    const validatedData = createWebhookSchema.safeParse(body);
    if (!validatedData.success) {
      throw new BadRequestException(validatedData.error.issues);
    }
    const data = validatedData.data;

    const secret = crypto.randomBytes(32).toString('hex');

    const webhook = await prisma.workspaceWebhook.create({
      data: {
        workspaceId: workspace.id,
        name: data.name,
        url: data.url,
        secret,
        events: data.events,
      },
    });

    return webhook;
  }

  @Patch(':webhookId')
  @ApiOperation({ summary: 'Update a webhook' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiParam({ name: 'webhookId', description: 'The webhook ID' })
  @ApiBody({ type: UpdateWorkspaceWebhookDto })
  @ApiResponse({ status: 200, description: 'Webhook updated' })
  async updateWebhook(
    @CurrentUser() user: User,
    @Param('slug') slug: string,
    @Param('webhookId') webhookId: string,
    @Body() body: UpdateWorkspaceWebhookDto
  ) {
    /**
     * ⚡ Performance Optimization:
     * 1. Consolidates workspace lookup and membership verification into a single database query.
     * 2. Reduces database round-trips from 2 down to 1.
     */
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    const member = workspace.members[0];

    if (!member || !['owner', 'admin'].includes(member.role)) {
      throw new ForbiddenException('Forbidden');
    }

    const validatedData = updateWebhookSchema.safeParse(body);
    if (!validatedData.success) {
      throw new BadRequestException(validatedData.error.issues);
    }
    const data = validatedData.data;

    /**
     * ⚡ Performance Optimization & Security Hardening (BOLA / IDOR Mitigation):
     * Replaces `prisma.workspaceWebhook.findFirst` with a direct O(1) primary key point lookup via `findUnique` on 'id'.
     * Workspace scoping (`existingWebhook.workspaceId === workspace.id`) is validated in application memory.
     * This leverages the primary key B-Tree index, avoiding multi-column index scans.
     */
    const existingWebhook = await prisma.workspaceWebhook.findUnique({
      where: { id: webhookId },
      select: { id: true, workspaceId: true },
    });

    if (!existingWebhook || existingWebhook.workspaceId !== workspace.id) {
      throw new NotFoundException('Webhook not found');
    }

    const webhook = await prisma.workspaceWebhook.update({
      where: { id: webhookId },
      data,
    });

    return webhook;
  }

  @Delete(':webhookId')
  @ApiOperation({ summary: 'Delete a webhook' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiParam({ name: 'webhookId', description: 'The webhook ID' })
  @ApiResponse({ status: 200, description: 'Webhook deleted' })
  async deleteWebhook(@CurrentUser() user: User, @Param('slug') slug: string, @Param('webhookId') webhookId: string) {
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    const member = workspace.members[0];

    if (!member || !['owner', 'admin'].includes(member.role)) {
      throw new ForbiddenException('Forbidden');
    }

    /**
     * 🛡️ Security Hardening (BOLA / IDOR Mitigation):
     * Perform an atomic deletion with workspace boundary validation (workspaceId: workspace.id).
     * Prevents cross-workspace deletion if an attacker provides a webhook ID belonging to another workspace.
     */
    const result = await prisma.workspaceWebhook.deleteMany({
      where: {
        id: webhookId,
        workspaceId: workspace.id,
      },
    });

    if (result.count === 0) {
      throw new NotFoundException('Webhook not found');
    }

    return { message: 'Webhook deleted successfully' };
  }
}