# apps/docs/src/content/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v3/workspaces/{slug}": {
      "get": {
        "description": "Retrieve details of a specific workspace by its slug.",
        "operationId": "V3WorkspacesController_getWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "workspace": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string",
                              "nullable": true
                            },
                            "icon": {
                              "type": "string",
                              "nullable": true
                            },
                            "industry": {
                              "type": "string",
                              "nullable": true
                            },
                            "brandingConfig": {
                              "type": "object",
                              "nullable": true
                            },
                            "createdAt": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Missing provisioning:workspaces scope or unauthorized workspace access."
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get workspace details (Enterprise M2M)",
        "tags": [
          "V3 Workspaces"
        ]
      },
      "patch": {
        "description": "Update the configuration and metadata of a specific workspace.",
        "operationId": "V3WorkspacesController_updateWorkspace",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V3UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "workspace": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string",
                              "nullable": true
                            },
                            "icon": {
                              "type": "string",
                              "nullable": true
                            },
                            "industry": {
                              "type": "string",
                              "nullable": true
                            },
                            "brandingConfig": {
                              "type": "object",
                              "nullable": true
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Missing provisioning:workspaces scope or unauthorized access."
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a workspace (Enterprise M2M)",
        "tags": [
          "V3 Workspaces"
        ]
      },
      "delete": {
        "description": "Permanently deletes a specific workspace.",
        "operationId": "V3WorkspacesController_deleteWorkspace",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Missing provisioning:workspaces scope or unauthorized access."
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a workspace (Enterprise M2M)",
        "tags": [
          "V3 Workspaces"
        ]
      }
    },
...
    "/api/v3/workspaces/{slug}/webhooks": {
      "get": {
        "description": "Requires webhooks:read scope. Retrieves all webhooks configured for this workspace.",
        "operationId": "V3WebhooksController_getWebhooks",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of webhooks returned successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all webhooks in the workspace",
        "tags": [
          "V3 Webhooks"
        ]
      },
      "post": {
        "description": "Requires webhooks:write scope.",
        "operationId": "V3WebhooksController_createWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V3CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new webhook",
        "tags": [
          "V3 Webhooks"
        ]
      }
    },
    "/api/v3/workspaces/{slug}/webhooks/{webhookId}": {
      "get": {
        "description": "Requires webhooks:read scope.",
        "operationId": "V3WebhooksController_getWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook details returned successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get details of a specific webhook",
        "tags": [
          "V3 Webhooks"
        ]
      },
      "patch": {
        "description": "Requires webhooks:write scope.",
        "operationId": "V3WebhooksController_updateWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V3UpdateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook updated successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a webhook",
        "tags": [
          "V3 Webhooks"
        ]
      },
      "delete": {
        "description": "Requires webhooks:write scope.",
        "operationId": "V3WebhooksController_deleteWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deleted successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a webhook",
        "tags": [
          "V3 Webhooks"
        ]
      }
    },
...
    "/api/integrations/huly/webhook/{id}": {
      "post": {
        "operationId": "IntegrationsController_handleHulyWebhook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Handle Huly webhook",
        "tags": [
          "Integrations"
        ]
      }
    },
...
    "/api/integrations/webhooks/{webhookId}": {
      "patch": {
        "operationId": "IntegrationsController_updateWebhook",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an integration webhook",
        "tags": [
          "Integrations"
        ]
      },
      "delete": {
        "operationId": "IntegrationsController_deleteWebhook",
        "parameters": [
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an integration webhook",
        "tags": [
          "Integrations"
        ]
      }
    },
...
    "/api/workspaces/{slug}": {
      "get": {
        "operationId": "WorkspacesController_getWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details"
          },
          "403": {
            "description": "Forbidden: Not a member"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get workspace details by slug",
        "tags": [
          "Workspaces"
        ]
      },
      "patch": {
        "operationId": "WorkspacesController_updateWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully"
          },
          "403": {
            "description": "Forbidden: Not an owner or admin"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workspace details",
        "tags": [
          "Workspaces"
        ]
      },
      "delete": {
        "operationId": "WorkspacesController_deleteWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully"
          },
          "403": {
            "description": "Forbidden: Only owner can delete"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a workspace",
        "tags": [
          "Workspaces"
        ]
      }
    },
...
    "/api/workspaces/{slug}/webhooks": {
      "get": {
        "operationId": "WebhooksController_getWebhooks",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of webhooks"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all webhooks for a workspace",
        "tags": [
          "Webhooks"
        ]
      },
      "post": {
        "operationId": "WebhooksController_createWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new webhook for a workspace",
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/workspaces/{slug}/webhooks/{webhookId}": {
      "patch": {
        "operationId": "WebhooksController_updateWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a webhook",
        "tags": [
          "Webhooks"
        ]
      },
      "delete": {
        "operationId": "WebhooksController_deleteWebhook",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "webhookId",
            "required": true,
            "in": "path",
            "description": "The webhook ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a webhook",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Scyrme Chat API",
    "description": "Comprehensive API documentation for Scyrme Chat V2 and V3 Enterprise. This documentation describes both standard V2 integrations and high-performance, secure V3 Enterprise M2M interfaces.",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "V3UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Corp Updated",
            "description": "The display name of the workspace"
          },
          "icon": {
            "type": "string",
            "example": "building",
            "description": "Icon identifier or URL"
          },
          "description": {
            "type": "string",
            "example": "Updated description for Acme Corp",
            "description": "A description for the workspace."
          },
          "industry": {
            "type": "string",
            "example": "Technology",
            "description": "The industry categorization of the workspace."
          },
          "brandingConfig": {
            "type": "object",
            "description": "Custom branding configuration"
          }
        }
      },
...
      "V3CreateWebhookDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Webhook"
          },
          "url": {
            "type": "string",
            "example": "https://example.com/webhook"
          },
          "events": {
            "example": [
              "message.sent",
              "channel.created"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "url",
          "events"
        ]
      },
      "V3UpdateWebhookDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Workspace Name"
          },
          "icon": {
            "type": "string",
            "example": "https://example.com/new-icon.png"
          },
          "banner": {
            "type": "string",
            "example": "https://example.com/banner.png"
          },
          "description": {
            "type": "string",
            "example": "Updated description"
          },
          "settings": {
            "type": "object"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "customDomain": {
            "type": "string"
          },
          "brandingConfig": {
            "type": "object"
          },
          "industry": {
            "type": "string"
          }
        }
      },
...
      "CreateWorkspaceWebhookDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Webhook"
          },
          "url": {
            "type": "string",
            "example": "https://example.com/webhook"
          },
          "events": {
            "example": [
              "message.sent",
              "channel.created"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "url",
          "events"
        ]
      },
      "UpdateWorkspaceWebhookDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    }
  }
}